services/NeuroNets_MM/requirements.txt
numpy~=1.19.2
Flask==2.0.2
opencv-python==4.5.4.58
tensorflow==2.6.0
keras==2.6.*
python-dotenv
